+++ CHANGELOG.md
@@ -2,7 +2,22 @@
 
 ## 2.0.0 (2023-??-??)
 
-Java 11 or later is required.
+Java 11 or later is required to run require-javadoc.
+
+## 1.0.9 (2024-03-28)
+
+Don't require documentation on record parameters (fields), because Javadoc
+requires `@param` tags in the Javadoc for the record.
+
+## 1.0.8 (2024-03-27)
+
+Bugfix: reinstate support for compiling and running under JDK 8.
+
+## 1.0.7 (2024-03-26)
+
+Support Java 17 syntax.
+
+Support compiling and running under JDK 21.
 
 ## 1.0.6 (2022-12-02)